+++ cluster-cm/src/main/java/com/sequenceiq/cloudbreak/cm/polling/ClouderaManagerCommandPollerObject.java
@@ -1,7 +1,5 @@
 package com.sequenceiq.cloudbreak.cm.polling;
 
-import java.math.BigDecimal;
-
 import com.cloudera.api.swagger.client.ApiClient;
 import com.sequenceiq.cloudbreak.domain.stack.Stack;
 
@@ -11,16 +9,16 @@ public class ClouderaManagerCommandPollerObject extends ClouderaManagerPollerObj
 
     private final ApiClient apiClient;
 
-    private final BigDecimal id;
+    private final Integer id;
 
-    public ClouderaManagerCommandPollerObject(Stack stack, ApiClient apiClient, BigDecimal id) {
+    public ClouderaManagerCommandPollerObject(Stack stack, ApiClient apiClient, Integer id) {
         super(stack, apiClient);
         this.stack = stack;
         this.apiClient = apiClient;
         this.id = id;
     }
 
-    public BigDecimal getId() {
+    public Integer getId() {
         return id;
     }
 }
